+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-#### Fixed
+### Changed
+- Allow REST API users endpoints if user is logged in and can edit_posts
+
+### Fixed
 - Escape outputting localization functions ask_e, asv_e and pll_e return
 - On `widgets_init` remove recent comments style